# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreatePATRequest:
    properties:
      name:
        type: string
    type: object
  handlers.CreateSSHKeyRequest:
    properties:
      name:
        type: string
      publicKey:
        type: string
    type: object
...
  handlers.CreateTrustedDomainRequest:
    properties:
      domain:
        type: string
    type: object
  handlers.CreateWorkspaceRequest:
    properties:
      name:
        type: string
    type: object
...
  handlers.UpdateProfileRequest:
    properties:
      email:
        type: string
      role:
        type: string
    type: object
...
  models.Workspace:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      ownerId:
        type: string
      preferredRegion:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Vessl API Documentation
  title: Vessl API
  version: "1.0"
paths:
...
  /api/profile:
    get:
      consumes:
      - application/json
      description: GetProfile endpoint
      produces:
      - application/json
      responses: {}
      summary: GetProfile endpoint
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: UpdateProfile endpoint
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses: {}
      summary: UpdateProfile endpoint
      tags:
      - Users
  /api/profile/tokens:
    get:
      consumes:
      - application/json
      description: ListPATs endpoint
      produces:
      - application/json
      responses: {}
      summary: ListPATs endpoint
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: CreatePAT endpoint
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePATRequest'
      produces:
      - application/json
      responses: {}
      summary: CreatePAT endpoint
      tags:
      - Users
  /api/profile/tokens/{id}:
    delete:
      consumes:
      - application/json
      description: DeletePAT endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: DeletePAT endpoint
      tags:
      - Users
...
  /api/teams/{teamId}/audit-logs:
    get:
      consumes:
      - application/json
      description: ListAuditLogs endpoint
      parameters:
      - description: teamId
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListAuditLogs endpoint
      tags:
      - Workspaces
...
  /api/teams/{teamId}/ssh-keys:
    get:
      consumes:
      - application/json
      description: ListSSHKeys endpoint
      parameters:
      - description: teamId
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListSSHKeys endpoint
      tags:
      - Workspaces
    post:
      consumes:
      - application/json
      description: CreateSSHKey endpoint
      parameters:
      - description: teamId
        in: path
        name: teamId
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSSHKeyRequest'
      produces:
      - application/json
      responses: {}
      summary: CreateSSHKey endpoint
      tags:
      - Workspaces
  /api/teams/{teamId}/trusted-domains:
    get:
      consumes:
      - application/json
      description: ListTrustedDomains endpoint
      parameters:
      - description: teamId
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListTrustedDomains endpoint
      tags:
      - Workspaces
    post:
      consumes:
      - application/json
      description: CreateTrustedDomain endpoint
      parameters:
      - description: teamId
        in: path
        name: teamId
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTrustedDomainRequest'
      produces:
      - application/json
      responses: {}
      summary: CreateTrustedDomain endpoint
      tags:
      - Workspaces
...
  /api/vercel/projects/{id}/env:
    get:
      consumes:
      - application/json
      description: GetProjectEnv endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: GetProjectEnv endpoint
      tags:
      - Projects
...
  /api/workspaces:
    post:
      consumes:
      - application/json
      description: Create endpoint
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWorkspaceRequest'
      produces:
      - application/json
      responses: {}
      summary: Create endpoint
      tags:
      - Workspaces
  /api/workspaces/{id}:
    delete:
      consumes:
      - application/json
      description: Delete endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete endpoint
      tags:
      - Workspaces
    get:
      consumes:
      - application/json
      description: Get endpoint
      parameters:
      - description: teamId
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get endpoint
      tags:
      - Workspaces
    put:
      consumes:
      - application/json
      description: Update endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Workspace'
      produces:
      - application/json
      responses: {}
      summary: Update endpoint
      tags:
      - Workspaces
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"